requirements.txt
Django==1.7.7
Markdown==2.6.1
django-filter==0.9.2
django-froala-editor==1.2.6
djangorestframework==3.1.1
gunicorn==19.3.0
pytz==2015.2
redis==2.10.3
wsgiref==0.1.2
